requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.4
feedparser==6.0.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-decouple==3.3
pytz==2020.5
schedule==0.6.0
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
